optimized position saving concept
