+++ Assets/Prefabs/Scene Transition/LoadWorlds.prefab
@@ -29,7 +29,8 @@ Transform:
   m_LocalPosition: {x: -45, y: 19, z: 0.13}
   m_LocalScale: {x: 15, y: 1.5, z: 1}
   m_ConstrainProportionsScale: 0
-  m_Children: []
+  m_Children:
+  - {fileID: 9126102115852185318}
   m_Father: {fileID: 0}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -45,11 +46,9 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 209e6f8ed06439f4b945e8fc2aed97a7, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  sceneLoadOne: World 1
-  sceneLoadTwo: World 3
-  sceneUnloadOne: World 2
-  sceneUnloadTwo: World 4
-  loadingFinished: 1
+  sceneOrigin: 
+  sceneDestination: 
+  sceneDestinationIndex: 0
 --- !u!61 &5562204860208180053
 BoxCollider2D:
   m_ObjectHideFlags: 0
@@ -76,3 +75,75 @@ BoxCollider2D:
   serializedVersion: 2
   m_Size: {x: 1, y: 2.303162}
   m_EdgeRadius: 0
+--- !u!1 &6835248662306171259
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 9126102115852185318}
+  - component: {fileID: 4830186715922675611}
+  - component: {fileID: 5758370617611294528}
+  m_Layer: 0
+  m_Name: UpdateCurrentArea
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &9126102115852185318
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6835248662306171259}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0.3724, y: -0.092, z: 0}
+  m_LocalScale: {x: 0.037296034, y: 0.047534164, z: 1}
+  m_ConstrainProportionsScale: 0
+  m_Children: []
+  m_Father: {fileID: 5562204860208180055}
+  m_RootOrder: 0
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!61 &4830186715922675611
+BoxCollider2D:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6835248662306171259}
+  m_Enabled: 1
+  m_Density: 1
+  m_Material: {fileID: 0}
+  m_IsTrigger: 1
+  m_UsedByEffector: 0
+  m_UsedByComposite: 0
+  m_Offset: {x: -0.12602401, y: 0.142622}
+  m_SpriteTilingProperty:
+    border: {x: 0, y: 0, z: 0, w: 0}
+    pivot: {x: 0, y: 0}
+    oldSize: {x: 0, y: 0}
+    newSize: {x: 0, y: 0}
+    adaptiveTilingThreshold: 0
+    drawMode: 0
+    adaptiveTiling: 0
+  m_AutoTiling: 0
+  serializedVersion: 2
+  m_Size: {x: 2.9209142, y: 3.5359802}
+  m_EdgeRadius: 0
+--- !u!114 &5758370617611294528
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6835248662306171259}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 79f662525069c5e40846ff28f4d17768, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  loadMapsTrigger: {fileID: 5562204860208180054}
