+++ Assets/Scripts/GameManager/DTOs/PlayersStatisticDTO.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 6cd39f27cde0f9a4191f1c7c6280d169
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
