--- Assets/Scripts/GameManager/DTOs/PlayerstatisticDTO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cbac556afcc6ac14099ca0e70e57e01e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
