+++ Assets/Scripts/Scene Loading/UpdateArea.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 79f662525069c5e40846ff28f4d17768
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
